requirements.txt
python-pptx
matplotlib~=3.1.2
comtypes~=1.1.7
pytest~=5.3.3
setuptools~=40.8.0